requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
billiard==3.6.2.0
boto3==1.11.15
botocore==1.14.15
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
jmespath==0.9.4
kombu==4.6.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.4
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
zipp==2.1.0
